experimentation/requirements.txt
## python==3.9
numpy==1.26.0
faiss-cpu==1.7.4
fastparquet==2024.5.0
torch==2.1.1
transformers==4.44.2

## might need to use conda to install faiss if doesnt work
## conda install -c pytorch faiss-cpu